requirements.txt
asgiref==3.8.1
Django==4.2
sqlparse==0.5.0
python-dotenv